+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Todo Tree Change Log
 
+- Make storage folder creation (if required) more robust
+
 ## v0.0.193 - 2020-12-31
 
 - Add support for a simple colour scheme to apply to the highlights